feat(nav): Phase 3 ActionChips — extract useActionBarModals, render chips in drawer (#257)
